Debug 21
I think I finally did it! It's not perfect, the collectables are still very very junkie with one of them not being implemented yet but I think it is a whole lot better when it first started though. I made changes to some of the icons and figured out how and they are not blurry anymore. change the text within ink files and figure out how to make a shopping button without just saying yes to it all the time! I think I will go back and change some more at a later day but I am happy with this for now!